Refactor builder module
